+++ include/ArduinoJson/Arduino/Print.hpp
@@ -20,6 +20,7 @@ class Print {
 
   size_t print(const char[]);
   size_t print(double, int = 2);
+  size_t print(int);
   size_t print(long);
   size_t println();
 };