+++ include/ArduinoJson/Internals/JsonFloat.hpp
@@ -1,0 +1,20 @@
+// Copyright Benoit Blanchon 2014-2015
+// MIT License
+//
+// Arduino JSON library
+// https://github.com/bblanchon/ArduinoJson
+
+#pragma once
+
+namespace ArduinoJson {
+namespace Internals {
+
+#ifdef ARDUINO
+// On embedded platform, we with use float instead of double to keep JsonVariant
+// small (issue #134)
+typedef float JsonFloat;
+#else
+typedef double JsonFloat;
+#endif
+}
+}