+++ include/ArduinoJson/Internals/JsonInteger.hpp
@@ -1,0 +1,13 @@
+// Copyright Benoit Blanchon 2014-2015
+// MIT License
+//
+// Arduino JSON library
+// https://github.com/bblanchon/ArduinoJson
+
+#pragma once
+
+namespace ArduinoJson {
+namespace Internals {
+typedef long JsonInteger;
+}
+}