+++ src/Arduino/Print.cpp
@@ -60,6 +60,12 @@ size_t Print::print(long value) {
   return print(tmp);
 }
 
+size_t Print::print(int value) {
+  char tmp[32];
+  sprintf(tmp, "%d", value);
+  return print(tmp);
+}
+
 size_t Print::println() { return write('\r') + write('\n'); }
 
 #endif